Move TestHttpProxyServer and supporting classes into an internal package
This way we can have some time to polish the API a bit before we decide to make
it public.
